#127 refactor: 테이블 코드 분리
