+++ src/components/Table/index.js
@@ -1,0 +1,3 @@
+export { default as AttendeeTable } from './AttendeeTable';
+
+export { default as TotalStatisticsPage } from './AttendeeTable';
