+++ src/components/index.js
@@ -12,3 +12,4 @@ export * from './Input';
 export * from './Textarea';
 export * from './TopNavigation';
 export * from './Sidebar';
+export * from './Table';
